rename test_operators to testutils
